+++ src/Services/Sale/SaleServiceBuilder.php
@@ -18,7 +18,15 @@ use Bitrix24\SDK\Core\Credentials\Scope;
 use Bitrix24\SDK\Services\AbstractServiceBuilder;
 use Bitrix24\SDK\Services\Sale\Property\Service\Property;
 use Bitrix24\SDK\Services\Sale\PropertyVariant\Service\PropertyVariant;
+use Bitrix24\SDK\Services\Sale\Status\Service\Status;
+use Bitrix24\SDK\Services\Sale\StatusLang\Service\StatusLang;
+use Bitrix24\SDK\Services\Sale\PersonTypeStatus\Service\PersonTypeStatus;
 
+/**
+ * Class SaleServiceBuilder
+ *
+ * @package Bitrix24\SDK\Services\Sale
+ */
 #[ApiServiceBuilderMetadata(new Scope(['sale']))]
 class SaleServiceBuilder extends AbstractServiceBuilder
 {
@@ -29,6 +37,30 @@ class SaleServiceBuilder extends AbstractServiceBuilder
     {
         if (!isset($this->serviceCache[__METHOD__])) {
             $this->serviceCache[__METHOD__] = new Property(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function status(): Status
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Status(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function personTypeStatus(): PersonTypeStatus
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new PersonTypeStatus(
                 $this->core,
                 $this->log
             );
@@ -49,13 +81,22 @@ class SaleServiceBuilder extends AbstractServiceBuilder
         return $this->serviceCache[__METHOD__];
     }
 
-    /**
-     * Order property variants service (sale.propertyvariant.*)
-     */
     public function propertyVariant(): PropertyVariant
     {
         if (!isset($this->serviceCache[__METHOD__])) {
             $this->serviceCache[__METHOD__] = new PropertyVariant(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function statusLang(): StatusLang
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new StatusLang(
                 $this->core,
                 $this->log
             );